neural types refactored
